[CHORE] Detail 부분 하나의 파일로 합치기 진행(#436)
