+++ Manito/Manito/Screens/Detail-Ing/DetailngCodebaseViewController.swift
@@ -12,8 +12,8 @@ import SnapKit
 // FIXME: 스토리보드 삭제 후 클래스명 변경 요
 final class DetailingCodebaseViewController: BaseViewController {
     
-    private let detailIngService: DetailIngAPI = DetailIngAPI(apiService: APIService())
-    private let detailDoneService: DetailDoneAPI = DetailDoneAPI(apiService: APIService())
+    private let detailIngService: DetailRoomAPI = DetailRoomAPI(apiService: APIService())
+    private let detailDoneService: DetailRoomAPI = DetailRoomAPI(apiService: APIService())
     
     private enum RoomType: String {
         case PROCESSING
